# src/test/resources/openapi/warehouse.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Warehouse API",
    "description" : "REST API системы управления складом",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/inventory/stocktake" : {
      "post" : {
        "tags" : [ "Инвентаризация" ],
        "summary" : "Провести инвентаризацию",
        "operationId" : "stocktake",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StocktakeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ValidationErrorResponse"
                  }, {
                    "$ref" : "#/components/schemas/ErrorResponse"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StockMovementResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/reports/stock-valuation" : {
      "get" : {
        "tags" : [ "Отчёты" ],
        "summary" : "Оценка складских запасов",
        "description" : "Σ quantity × cost с разрезом по категориям. Доступно только ADMIN.",
        "operationId" : "getStockValuation",
        "responses" : {
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ValidationErrorResponse"
                  }, {
                    "$ref" : "#/components/schemas/ErrorResponse"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StockValuationResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/reports/low-stock" : {
      "get" : {
        "tags" : [ "Отчёты" ],
        "summary" : "Товары ниже минимального остатка",
        "operationId" : "getLowStock",
        "responses" : {
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ValidationErrorResponse"
                  }, {
                    "$ref" : "#/components/schemas/ErrorResponse"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LowStockReportResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/reports/expiring" : {
      "get" : {
        "tags" : [ "Отчёты" ],
        "summary" : "Партии с истекающим сроком годности",
        "description" : "Найти партии, срок годности которых истекает в ближайшие N дней. FEFO: first-expire-first-out. Доступно только ADMIN.",
        "operationId" : "getExpiringBatches",
        "parameters" : [ {
          "name" : "days",
          "in" : "query",
          "required" : true,
          "schema" : {
            "maximum" : 365,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ValidationErrorResponse"
                  }, {
                    "$ref" : "#/components/schemas/ErrorResponse"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpiringBatchesReport"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "required" : [ "error", "message", "timestamp" ],
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ValidationErrorResponse" : {
        "required" : [ "error", "fields" ],
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "fields" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Сообщения валидации по имени поля, например {\"sku\": \"must not be blank\"}"
          }
        }
      },
...
      "StockMovementResponse" : {
        "required" : [ "batchId", "createdAt", "expiryDate", "itemId", "lowStockAlert", "movementId", "quantity", "stockAfter", "transferId", "type", "warehouseId", "warehouseName" ],
        "type" : "object",
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "movementId" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "type" : {
            "type" : "string",
            "nullable" : true,
            "enum" : [ "RECEIVE", "WRITE_OFF", "EXPIRED", "ADJUSTMENT", "TRANSFER_OUT", "TRANSFER_IN" ]
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "stockAfter" : {
            "type" : "integer",
            "format" : "int32"
          },
          "batchId" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "expiryDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "lowStockAlert" : {
            "type" : "boolean"
          },
          "warehouseId" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "warehouseName" : {
            "type" : "string",
            "nullable" : true
          },
          "transferId" : {
            "type" : "string",
            "format" : "uuid",
            "nullable" : true
          }
        }
      },
...
      "StocktakeRequest" : {
        "required" : [ "countedQuantity", "itemId" ],
        "type" : "object",
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "countedQuantity" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "surplusExpiryDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CategoryValuation" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "valuation" : {
            "type" : "number"
          }
        }
      },
      "StockValuationResponse" : {
        "type" : "object",
        "properties" : {
          "totalValuation" : {
            "type" : "number"
          },
          "byCategory" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CategoryValuation"
            }
          }
        }
      },
      "LowStockItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sku" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "currentStock" : {
            "type" : "integer",
            "format" : "int64"
          },
          "minStock" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deficit" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "LowStockReportResponse" : {
        "type" : "object",
        "properties" : {
          "generatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LowStockItem"
            }
          }
        }
      },
      "ExpiringBatch" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sku" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "warehouseId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "warehouseName" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expiryDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ExpiringBatchesReport" : {
        "type" : "object",
        "properties" : {
          "generatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpiringBatch"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}